Fixed weird administrator issue
Yesterday, I added the flappy bird sprite sheet to the project, but, when I tried to change aspects about how the sprite sheet was read, it would revert back to the default. I managed to solve what seemed to be the first part of the problem - that I was always forced to enter Unity in administrator. I fixed it by changing my user account control settings to "always notify".

Now I can enter Unity as a standard user, but the main issue still persists, which I'm going to try to fix now. It seems I'm still getting the same errors from before, "TextureImporterInspector.OnInspectorGUI must call ApplyRevertGUI to avoid unexpected behaviour." and "Could not find the requested Platform Texture Settings. This is incorrect, did initialization fail?".

It seems that the former is where the problem lies, as it's there when I open the program and the latter just shows up when I try to change things with the spritesheet. Anyway, time to try and fix that so I can actually use sprites.