Not Good
Not much progress was made. Found out I can still use the sprites and run the game with them, just not change them, so if it comes to it I guess I can do that.

I'm not sure if this is part of the problem, but pretty much all of the 2D extensions except 2D common are locked and every time I unlock them, they lock themselves again. I looked up how to stop it and I'm supposed to change enableLockFile in the manifest.json file to false, but I couldn't even find that property there and it didn't look like the file had properties like that where it's a place I could add enableLockFile to, it had a bunch of com. things. Maybe there's another manifest file or something. I'll find it some other time or something.

Also I removed versioncontrol, I thought that might help with the whole locking thing, but it didn't so maybe I should reinstall it.