Issue Fixed! Flappy bird Sprite set-up complete!
It turned out the fix was actually really easy, it was the first forum post online, I just had to change buildTarget's value to "Server" in the metadata file of the spritesheet! Errors gone!

I was able to change the aspects of the sprites. Then, I changed the aspect ratio of the game and added the background, pipes and flappy bird. I was also able to animate the flappy bird. Hopefully everything is mostly right, I couldn't remember what was to be done in class.